dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    ./compile-requirements
#
appdirs==1.4.4            # via black, virtualenv
attrs==20.3.0             # via -r ./py-deploy/requirements.txt, black, jsonschema, pytest
base58==2.1.0             # via -r ./py-deploy/requirements.txt, multiaddr
bitarray==1.2.2           # via -r ./py-deploy/requirements.txt, eth-account
black==20.8b0             # via -r dev-requirements.in
blake2b-py==0.1.3         # via -r ./py-deploy/requirements.txt, py-evm
cached-property==1.5.2    # via -r ./py-deploy/requirements.txt, py-ecc, py-evm
certifi==2020.12.5        # via -r ./py-deploy/requirements.txt, requests
cfgv==3.2.0               # via pre-commit
chardet==4.0.0            # via -r ./py-deploy/requirements.txt, requests
click==7.1.2              # via -r ./py-deploy/requirements.txt, black, contract-deploy-tools, eth-tester-rpc, pip-tools
contract-deploy-tools==0.9.0  # via -r ./py-deploy/requirements.txt, -r dev-requirements.in
cytoolz==0.11.0           # via -r ./py-deploy/requirements.txt, eth-keyfile, eth-tester-rpc, eth-utils
dataclasses==0.8          # via black
distlib==0.3.1            # via virtualenv
eth-abi==2.1.1            # via -r ./py-deploy/requirements.txt, eth-account, eth-tester, web3
eth-account==0.5.4        # via -r ./py-deploy/requirements.txt, web3
eth-bloom==1.0.3          # via -r ./py-deploy/requirements.txt, py-evm
eth-hash[pycryptodome,pysha3]==0.2.0  # via -r ./py-deploy/requirements.txt, eth-bloom, eth-tester, eth-tester-rpc, eth-utils, trie, web3
eth-keyfile==0.5.1        # via -r ./py-deploy/requirements.txt, contract-deploy-tools, eth-account
eth-keys==0.3.3           # via -r ./py-deploy/requirements.txt, eth-account, eth-keyfile, eth-tester, py-evm
eth-rlp==0.2.1            # via -r ./py-deploy/requirements.txt, eth-account
eth-tester-rpc==0.5.0b1   # via -r ./py-deploy/requirements.txt, contract-deploy-tools
eth-tester[py-evm]==0.5.0b3  # via -r ./py-deploy/requirements.txt, contract-deploy-tools, eth-tester-rpc
eth-typing==2.2.2         # via -r ./py-deploy/requirements.txt, eth-abi, eth-keys, eth-utils, py-ecc, py-evm, web3
eth-utils==1.9.5          # via -r ./py-deploy/requirements.txt, contract-deploy-tools, eth-abi, eth-account, eth-keyfile, eth-keys, eth-rlp, eth-tester, eth-tester-rpc, py-ecc, py-evm, rlp, trie, web3
filelock==3.0.12          # via virtualenv
flake8==3.8.4             # via -r dev-requirements.in
hexbytes==0.2.1           # via -r ./py-deploy/requirements.txt, eth-account, eth-rlp, trie, web3
identify==1.5.12          # via pre-commit
idna==2.10                # via -r ./py-deploy/requirements.txt, requests
importlib-metadata==3.4.0  # via -r ./py-deploy/requirements.txt, flake8, jsonschema, pluggy, pre-commit, pytest, virtualenv
importlib-resources==5.0.0  # via -r ./py-deploy/requirements.txt, netaddr, pre-commit, virtualenv
iniconfig==1.1.1          # via pytest
ipfshttpclient==0.7.0a1   # via -r ./py-deploy/requirements.txt, web3
json-rpc==1.13.0          # via -r ./py-deploy/requirements.txt, eth-tester-rpc
jsonschema==3.2.0         # via -r ./py-deploy/requirements.txt, web3
lru-dict==1.1.6           # via -r ./py-deploy/requirements.txt, py-evm, web3
mccabe==0.6.1             # via flake8
multiaddr==0.0.9          # via -r ./py-deploy/requirements.txt, ipfshttpclient
mypy-extensions==0.4.3    # via -r ./py-deploy/requirements.txt, black, mypy, py-ecc, py-evm
mypy==0.790               # via -r dev-requirements.in
netaddr==0.8.0            # via -r ./py-deploy/requirements.txt, multiaddr
nodeenv==1.5.0            # via pre-commit
packaging==20.8           # via pytest
parsimonious==0.8.1       # via -r ./py-deploy/requirements.txt, eth-abi
pathspec==0.8.1           # via black
pendulum==2.1.2           # via -r ./py-deploy/requirements.txt
pip-tools==5.5.0          # via -r dev-requirements.in
pluggy==0.13.1            # via pytest
pre-commit==2.9.3         # via -r dev-requirements.in
protobuf==3.14.0          # via -r ./py-deploy/requirements.txt, web3
py-ecc==4.1.0             # via -r ./py-deploy/requirements.txt, py-evm
py-evm==0.3.0a20          # via -r ./py-deploy/requirements.txt, eth-tester
py-solc==3.2.0            # via -r ./py-deploy/requirements.txt, contract-deploy-tools
py==1.10.0                # via pytest
pycodestyle==2.6.0        # via flake8
pycryptodome==3.9.9       # via -r ./py-deploy/requirements.txt, eth-hash, eth-keyfile
pyethash==0.1.27          # via -r ./py-deploy/requirements.txt, py-evm
pyflakes==2.2.0           # via flake8
pyparsing==2.4.7          # via packaging
pyrsistent==0.17.3        # via -r ./py-deploy/requirements.txt, jsonschema
pysha3==1.0.2             # via -r ./py-deploy/requirements.txt, eth-hash
pytest==6.2.1             # via -r dev-requirements.in
python-dateutil==2.8.1    # via -r ./py-deploy/requirements.txt, pendulum
pytzdata==2020.1          # via -r ./py-deploy/requirements.txt, pendulum
pyyaml==5.3.1             # via pre-commit
regex==2020.11.13         # via black
requests==2.25.1          # via -r ./py-deploy/requirements.txt, ipfshttpclient, web3
rlp==2.0.1                # via -r ./py-deploy/requirements.txt, eth-account, eth-rlp, eth-tester, py-evm, trie
semantic-version==2.8.5   # via -r ./py-deploy/requirements.txt, eth-tester, py-solc
setuptools-scm==5.0.1     # via -r dev-requirements.in
six==1.15.0               # via -r ./py-deploy/requirements.txt, jsonschema, multiaddr, parsimonious, protobuf, python-dateutil, virtualenv
sortedcontainers==2.3.0   # via -r ./py-deploy/requirements.txt, trie
texttable==1.6.3          # via -r dev-requirements.in
toml==0.10.2              # via black, pre-commit, pytest
toolz==0.11.1             # via -r ./py-deploy/requirements.txt, cytoolz
trie==2.0.0-alpha.5       # via -r ./py-deploy/requirements.txt, py-evm
trustlines-contracts-bin==1.1.7  # via -r ./py-deploy/requirements.txt
typed-ast==1.4.2          # via black, mypy
typing-extensions==3.7.4.3  # via -r ./py-deploy/requirements.txt, black, importlib-metadata, mypy, trie, web3
urllib3==1.26.2           # via -r ./py-deploy/requirements.txt, requests
varint==1.0.2             # via -r ./py-deploy/requirements.txt, multiaddr
virtualenv==20.3.1        # via pre-commit
web3==5.14.0              # via -r ./py-deploy/requirements.txt, contract-deploy-tools
websockets==8.1           # via -r ./py-deploy/requirements.txt, web3
werkzeug==1.0.1           # via -r ./py-deploy/requirements.txt, eth-tester-rpc
zipp==3.4.0               # via -r ./py-deploy/requirements.txt, importlib-metadata, importlib-resources

# The following packages are considered to be unsafe in a requirements file:
pip==20.3.3               # via -r dev-requirements.in, pip-tools
setuptools==51.1.2        # via -r dev-requirements.in, jsonschema, setuptools-scm
